feat: DataLoader & useListaAulas
